Splitted udf communication tests into unit tests and integration tests
