+++ tests/integration_tests/without_db/udf_communication/peer_communication/test_add_peer.py
@@ -1,24 +1,31 @@
+import sys
 import time
+import traceback
 from pathlib import Path
 from typing import Dict, List
 
 import pytest
 import structlog
 import zmq
+from numpy.random import RandomState
 from structlog import WriteLoggerFactory
 from structlog.tracebacks import ExceptionDictTransformer
 from structlog.types import FilteringBoundLogger
 
 from exasol_advanced_analytics_framework.udf_communication.connection_info import ConnectionInfo
-from exasol_advanced_analytics_framework.udf_communication.discovery import localhost
-from exasol_advanced_analytics_framework.udf_communication.discovery.localhost.communicator import \
-    CommunicatorFactory
-from exasol_advanced_analytics_framework.udf_communication.ip_address import Port, IPAddress
+from exasol_advanced_analytics_framework.udf_communication.ip_address import IPAddress
 from exasol_advanced_analytics_framework.udf_communication.peer import Peer
+from exasol_advanced_analytics_framework.udf_communication.peer_communicator import PeerCommunicator
+from exasol_advanced_analytics_framework.udf_communication.peer_communicator.forward_register_peer_config import \
+    ForwardRegisterPeerConfig
 from exasol_advanced_analytics_framework.udf_communication.peer_communicator.peer_communicator import key_for_peer
+from exasol_advanced_analytics_framework.udf_communication.socket_factory.fault_injection import \
+    FaultInjectionSocketFactory
+from exasol_advanced_analytics_framework.udf_communication.peer_communicator.peer_communicator_config import \
+    PeerCommunicatorConfig
 from exasol_advanced_analytics_framework.udf_communication.socket_factory.zmq_wrapper import ZMQSocketFactory
-from tests.udf_communication.peer_communication.conditional_method_dropper import ConditionalMethodDropper
-from tests.udf_communication.peer_communication.utils import TestProcess, BidirectionalQueue, assert_processes_finish, \
+from tests.integration_tests.without_db.udf_communication.peer_communication.conditional_method_dropper import ConditionalMethodDropper
+from tests.integration_tests.without_db.udf_communication.peer_communication.utils import TestProcess, BidirectionalQueue, assert_processes_finish, \
     PeerCommunicatorTestProcessParameter
 
 structlog.configure(
@@ -36,29 +43,48 @@ structlog.configure(
     ]
 )
 
-LOGGER: FilteringBoundLogger = structlog.get_logger(__name__)
+LOGGER: FilteringBoundLogger = structlog.get_logger()
 
 
 def run(parameter: PeerCommunicatorTestProcessParameter, queue: BidirectionalQueue):
-    discovery_port = Port(port=44444)
-    listen_ip = IPAddress(ip_address="127.1.0.1")
-    context = zmq.Context()
-    socket_factory = ZMQSocketFactory(context)
-    discovery_socket_factory = localhost.DiscoverySocketFactory()
-    peer_communicator = CommunicatorFactory().create(
-        group_identifier=parameter.group_identifier,
-        name=parameter.instance_name,
-        number_of_instances=parameter.number_of_instances,
-        listen_ip=listen_ip,
-        discovery_port=discovery_port,
-        socket_factory=socket_factory,
-        discovery_socket_factory=discovery_socket_factory)
-    queue.put(peer_communicator.my_connection_info)
-    if peer_communicator.are_all_peers_connected():
-        peers = peer_communicator.peers()
-        queue.put(peers)
-    else:
+    logger = LOGGER.bind(group_identifier=parameter.group_identifier, name=parameter.instance_name)
+    try:
+        listen_ip = IPAddress(ip_address=f"127.1.0.1")
+        context = zmq.Context()
+        socket_factory = ZMQSocketFactory(context)
+        socket_factory = FaultInjectionSocketFactory(socket_factory, 0.01, RandomState(parameter.seed))
+        com = PeerCommunicator(
+            name=parameter.instance_name,
+            number_of_peers=parameter.number_of_instances,
+            listen_ip=listen_ip,
+            group_identifier=parameter.group_identifier,
+            socket_factory=socket_factory,
+            config=PeerCommunicatorConfig(
+                forward_register_peer_config=ForwardRegisterPeerConfig(
+                    is_leader=False,
+                    is_enabled=False
+                )
+            ),
+        )
+        try:
+            queue.put(com.my_connection_info)
+            peer_connection_infos = queue.get()
+            for index, connection_info in peer_connection_infos.items():
+                com.register_peer(connection_info)
+            peers = com.peers(timeout_in_milliseconds=None)
+            logger.info("peers", number_of_peers=len(peers))
+            queue.put(peers)
+        finally:
+            com.close()
+            logger.info("after close")
+            context.destroy(linger=0)
+            logger.info("after destroy")
+            for frame in sys._current_frames().values():
+                stacktrace = traceback.format_stack(frame)
+                logger.info("Frame", stacktrace=stacktrace)
+    except Exception as e:
         queue.put([])
+        logger.exception("Exception during test")
 
 
 @pytest.mark.parametrize("number_of_instances, repetitions", [(2, 1000), (10, 100)])
@@ -66,7 +92,7 @@ def test_reliability(number_of_instances: int, repetitions: int):
     run_test_with_repetitions(number_of_instances, repetitions)
 
 
-REPETITIONS_FOR_FUNCTIONALITY = 2
+REPETITIONS_FOR_FUNCTIONALITY = 1
 
 
 def test_functionality_2():
@@ -89,7 +115,7 @@ def run_test_with_repetitions(number_of_instances: int, repetitions: int):
                     number_of_instances=number_of_instances)
         start_time = time.monotonic()
         group = f"{time.monotonic_ns()}"
-        expected_peers_of_threads, peers_of_threads = run_test(group, number_of_instances)
+        expected_peers_of_threads, peers_of_threads = run_test(group, number_of_instances, seed=i)
         assert expected_peers_of_threads == peers_of_threads
         end_time = time.monotonic()
         LOGGER.info(f"Finish iteration",
@@ -99,20 +125,21 @@ def run_test_with_repetitions(number_of_instances: int, repetitions: int):
                     duration=end_time - start_time)
 
 
-def run_test(group: str, number_of_instances: int):
+def run_test(group: str, number_of_instances: int, seed: int):
     connection_infos: Dict[int, ConnectionInfo] = {}
     parameters = [
         PeerCommunicatorTestProcessParameter(
             instance_name=f"i{i}", group_identifier=group,
             number_of_instances=number_of_instances,
-            seed=0)
+            seed=seed + i)
         for i in range(number_of_instances)]
     processes: List[TestProcess[PeerCommunicatorTestProcessParameter]] = \
         [TestProcess(parameter, run=run) for parameter in parameters]
     for i in range(number_of_instances):
         processes[i].start()
-    for i in range(number_of_instances):
         connection_infos[i] = processes[i].get()
+    for i in range(number_of_instances):
+        t = processes[i].put(connection_infos)
     assert_processes_finish(processes, timeout_in_seconds=180)
     peers_of_threads: Dict[int, List[ConnectionInfo]] = {}
     for i in range(number_of_instances):
