+++ tests/integration_tests/without_db/udf_communication/peer_communication/test_send_recv.py
@@ -1,16 +1,13 @@
-import sys
 import time
-import traceback
 from pathlib import Path
-from typing import Dict, List
+from typing import Dict, Set, List
 
 import pytest
 import structlog
 import zmq
-from numpy.random import RandomState
 from structlog import WriteLoggerFactory
 from structlog.tracebacks import ExceptionDictTransformer
-from structlog.types import FilteringBoundLogger
+from structlog.typing import FilteringBoundLogger
 
 from exasol_advanced_analytics_framework.udf_communication.connection_info import ConnectionInfo
 from exasol_advanced_analytics_framework.udf_communication.ip_address import IPAddress
@@ -18,14 +15,11 @@ from exasol_advanced_analytics_framework.udf_communication.peer import Peer
 from exasol_advanced_analytics_framework.udf_communication.peer_communicator import PeerCommunicator
 from exasol_advanced_analytics_framework.udf_communication.peer_communicator.forward_register_peer_config import \
     ForwardRegisterPeerConfig
-from exasol_advanced_analytics_framework.udf_communication.peer_communicator.peer_communicator import key_for_peer
-from exasol_advanced_analytics_framework.udf_communication.socket_factory.fault_injection import \
-    FaultInjectionSocketFactory
+from exasol_advanced_analytics_framework.udf_communication.socket_factory.zmq_wrapper import ZMQSocketFactory
 from exasol_advanced_analytics_framework.udf_communication.peer_communicator.peer_communicator_config import \
     PeerCommunicatorConfig
-from exasol_advanced_analytics_framework.udf_communication.socket_factory.zmq_wrapper import ZMQSocketFactory
-from tests.udf_communication.peer_communication.conditional_method_dropper import ConditionalMethodDropper
-from tests.udf_communication.peer_communication.utils import TestProcess, BidirectionalQueue, assert_processes_finish, \
+from tests.integration_tests.without_db.udf_communication.peer_communication.conditional_method_dropper import ConditionalMethodDropper
+from tests.integration_tests.without_db.udf_communication.peer_communication.utils import TestProcess, BidirectionalQueue, assert_processes_finish, \
     PeerCommunicatorTestProcessParameter
 
 structlog.configure(
@@ -47,44 +41,37 @@ LOGGER: FilteringBoundLogger = structlog.get_logger()
 
 
 def run(parameter: PeerCommunicatorTestProcessParameter, queue: BidirectionalQueue):
-    logger = LOGGER.bind(group_identifier=parameter.group_identifier, name=parameter.instance_name)
-    try:
-        listen_ip = IPAddress(ip_address=f"127.1.0.1")
-        context = zmq.Context()
-        socket_factory = ZMQSocketFactory(context)
-        socket_factory = FaultInjectionSocketFactory(socket_factory, 0.01, RandomState(parameter.seed))
-        com = PeerCommunicator(
-            name=parameter.instance_name,
-            number_of_peers=parameter.number_of_instances,
-            listen_ip=listen_ip,
-            group_identifier=parameter.group_identifier,
-            socket_factory=socket_factory,
-            config=PeerCommunicatorConfig(
-                forward_register_peer_config=ForwardRegisterPeerConfig(
-                    is_leader=False,
-                    is_enabled=False
-                )
+    listen_ip = IPAddress(ip_address=f"127.1.0.1")
+    context = zmq.Context()
+    socker_factory = ZMQSocketFactory(context)
+    com = PeerCommunicator(
+        name=parameter.instance_name,
+        number_of_peers=parameter.number_of_instances,
+        listen_ip=listen_ip,
+        group_identifier=parameter.group_identifier,
+        socket_factory=socker_factory,
+        config=PeerCommunicatorConfig(
+            forward_register_peer_config=ForwardRegisterPeerConfig(
+                is_leader=False,
+                is_enabled=False
             ),
-        )
-        try:
-            queue.put(com.my_connection_info)
-            peer_connection_infos = queue.get()
-            for index, connection_info in peer_connection_infos.items():
-                com.register_peer(connection_info)
-            peers = com.peers(timeout_in_milliseconds=None)
-            logger.info("peers", number_of_peers=len(peers))
-            queue.put(peers)
-        finally:
-            com.close()
-            logger.info("after close")
-            context.destroy(linger=0)
-            logger.info("after destroy")
-            for frame in sys._current_frames().values():
-                stacktrace = traceback.format_stack(frame)
-                logger.info("Frame", stacktrace=stacktrace)
-    except Exception as e:
-        queue.put([])
-        logger.exception("Exception during test")
+        ),
+    )
+    queue.put(com.my_connection_info)
+    peer_connection_infos = queue.get()
+    for index, connection_infos in peer_connection_infos.items():
+        com.register_peer(connection_infos)
+    com.wait_for_peers()
+    LOGGER.info("Peer is ready", name=parameter.instance_name)
+    for peer in com.peers():
+        if peer != Peer(connection_info=com.my_connection_info):
+            com.send(peer, [socker_factory.create_frame(parameter.instance_name.encode("utf8"))])
+    received_values: Set[str] = set()
+    for peer in com.peers():
+        if peer != Peer(connection_info=com.my_connection_info):
+            value = com.recv(peer)
+            received_values.add(value[0].to_bytes().decode("utf8"))
+    queue.put(received_values)
 
 
 @pytest.mark.parametrize("number_of_instances, repetitions", [(2, 1000), (10, 100)])
@@ -92,7 +79,7 @@ def test_reliability(number_of_instances: int, repetitions: int):
     run_test_with_repetitions(number_of_instances, repetitions)
 
 
-REPETITIONS_FOR_FUNCTIONALITY = 1
+REPETITIONS_FOR_FUNCTIONALITY = 2
 
 
 def test_functionality_2():
@@ -115,7 +102,7 @@ def run_test_with_repetitions(number_of_instances: int, repetitions: int):
                     number_of_instances=number_of_instances)
         start_time = time.monotonic()
         group = f"{time.monotonic_ns()}"
-        expected_peers_of_threads, peers_of_threads = run_test(group, number_of_instances, seed=i)
+        expected_peers_of_threads, peers_of_threads = run_test(group, number_of_instances)
         assert expected_peers_of_threads == peers_of_threads
         end_time = time.monotonic()
         LOGGER.info(f"Finish iteration",
@@ -125,30 +112,32 @@ def run_test_with_repetitions(number_of_instances: int, repetitions: int):
                     duration=end_time - start_time)
 
 
-def run_test(group: str, number_of_instances: int, seed: int):
+def run_test(group: str, number_of_instances: int):
     connection_infos: Dict[int, ConnectionInfo] = {}
     parameters = [
         PeerCommunicatorTestProcessParameter(
             instance_name=f"i{i}", group_identifier=group,
             number_of_instances=number_of_instances,
-            seed=seed + i)
+            seed=0)
         for i in range(number_of_instances)]
     processes: List[TestProcess[PeerCommunicatorTestProcessParameter]] = \
         [TestProcess(parameter, run=run) for parameter in parameters]
     for i in range(number_of_instances):
         processes[i].start()
+    for i in range(number_of_instances):
         connection_infos[i] = processes[i].get()
     for i in range(number_of_instances):
         t = processes[i].put(connection_infos)
     assert_processes_finish(processes, timeout_in_seconds=180)
-    peers_of_threads: Dict[int, List[ConnectionInfo]] = {}
+    received_values: Dict[int, Set[str]] = {}
     for i in range(number_of_instances):
-        peers_of_threads[i] = processes[i].get()
-    expected_peers_of_threads = {
-        i: sorted([
-            Peer(connection_info=connection_info)
-            for index, connection_info in connection_infos.items()
-        ], key=key_for_peer)
+        received_values[i] = processes[i].get()
+    expected_received_values = {
+        i: {
+            thread.parameter.instance_name
+            for index, thread in enumerate(processes)
+            if index != i
+        }
         for i in range(number_of_instances)
     }
-    return expected_peers_of_threads, peers_of_threads
+    return expected_received_values, received_values
