+++ tests/integration_tests/without_db/udf_communication/test_localhost_discovery.py
@@ -1,25 +1,24 @@
 import time
 from pathlib import Path
-from typing import Dict, Set, List
+from typing import Dict, List
 
 import pytest
 import structlog
 import zmq
 from structlog import WriteLoggerFactory
 from structlog.tracebacks import ExceptionDictTransformer
-from structlog.typing import FilteringBoundLogger
+from structlog.types import FilteringBoundLogger
 
 from exasol_advanced_analytics_framework.udf_communication.connection_info import ConnectionInfo
-from exasol_advanced_analytics_framework.udf_communication.ip_address import IPAddress
+from exasol_advanced_analytics_framework.udf_communication.discovery import localhost
+from exasol_advanced_analytics_framework.udf_communication.discovery.localhost.communicator import \
+    CommunicatorFactory
+from exasol_advanced_analytics_framework.udf_communication.ip_address import Port, IPAddress
 from exasol_advanced_analytics_framework.udf_communication.peer import Peer
-from exasol_advanced_analytics_framework.udf_communication.peer_communicator import PeerCommunicator
-from exasol_advanced_analytics_framework.udf_communication.peer_communicator.forward_register_peer_config import \
-    ForwardRegisterPeerConfig
+from exasol_advanced_analytics_framework.udf_communication.peer_communicator.peer_communicator import key_for_peer
 from exasol_advanced_analytics_framework.udf_communication.socket_factory.zmq_wrapper import ZMQSocketFactory
-from exasol_advanced_analytics_framework.udf_communication.peer_communicator.peer_communicator_config import \
-    PeerCommunicatorConfig
-from tests.udf_communication.peer_communication.conditional_method_dropper import ConditionalMethodDropper
-from tests.udf_communication.peer_communication.utils import TestProcess, BidirectionalQueue, assert_processes_finish, \
+from tests.integration_tests.without_db.udf_communication.peer_communication.conditional_method_dropper import ConditionalMethodDropper
+from tests.integration_tests.without_db.udf_communication.peer_communication.utils import TestProcess, BidirectionalQueue, assert_processes_finish, \
     PeerCommunicatorTestProcessParameter
 
 structlog.configure(
@@ -37,41 +36,29 @@ structlog.configure(
     ]
 )
 
-LOGGER: FilteringBoundLogger = structlog.get_logger()
+LOGGER: FilteringBoundLogger = structlog.get_logger(__name__)
 
 
 def run(parameter: PeerCommunicatorTestProcessParameter, queue: BidirectionalQueue):
-    listen_ip = IPAddress(ip_address=f"127.1.0.1")
+    discovery_port = Port(port=44444)
+    listen_ip = IPAddress(ip_address="127.1.0.1")
     context = zmq.Context()
-    socker_factory = ZMQSocketFactory(context)
-    com = PeerCommunicator(
-        name=parameter.instance_name,
-        number_of_peers=parameter.number_of_instances,
-        listen_ip=listen_ip,
+    socket_factory = ZMQSocketFactory(context)
+    discovery_socket_factory = localhost.DiscoverySocketFactory()
+    peer_communicator = CommunicatorFactory().create(
         group_identifier=parameter.group_identifier,
-        socket_factory=socker_factory,
-        config=PeerCommunicatorConfig(
-            forward_register_peer_config=ForwardRegisterPeerConfig(
-                is_leader=False,
-                is_enabled=False
-            ),
-        ),
-    )
-    queue.put(com.my_connection_info)
-    peer_connection_infos = queue.get()
-    for index, connection_infos in peer_connection_infos.items():
-        com.register_peer(connection_infos)
-    com.wait_for_peers()
-    LOGGER.info("Peer is ready", name=parameter.instance_name)
-    for peer in com.peers():
-        if peer != Peer(connection_info=com.my_connection_info):
-            com.send(peer, [socker_factory.create_frame(parameter.instance_name.encode("utf8"))])
-    received_values: Set[str] = set()
-    for peer in com.peers():
-        if peer != Peer(connection_info=com.my_connection_info):
-            value = com.recv(peer)
-            received_values.add(value[0].to_bytes().decode("utf8"))
-    queue.put(received_values)
+        name=parameter.instance_name,
+        number_of_instances=parameter.number_of_instances,
+        listen_ip=listen_ip,
+        discovery_port=discovery_port,
+        socket_factory=socket_factory,
+        discovery_socket_factory=discovery_socket_factory)
+    queue.put(peer_communicator.my_connection_info)
+    if peer_communicator.are_all_peers_connected():
+        peers = peer_communicator.peers()
+        queue.put(peers)
+    else:
+        queue.put([])
 
 
 @pytest.mark.parametrize("number_of_instances, repetitions", [(2, 1000), (10, 100)])
@@ -126,18 +113,15 @@ def run_test(group: str, number_of_instances: int):
         processes[i].start()
     for i in range(number_of_instances):
         connection_infos[i] = processes[i].get()
-    for i in range(number_of_instances):
-        t = processes[i].put(connection_infos)
     assert_processes_finish(processes, timeout_in_seconds=180)
-    received_values: Dict[int, Set[str]] = {}
+    peers_of_threads: Dict[int, List[ConnectionInfo]] = {}
     for i in range(number_of_instances):
-        received_values[i] = processes[i].get()
-    expected_received_values = {
-        i: {
-            thread.parameter.instance_name
-            for index, thread in enumerate(processes)
-            if index != i
-        }
+        peers_of_threads[i] = processes[i].get()
+    expected_peers_of_threads = {
+        i: sorted([
+            Peer(connection_info=connection_info)
+            for index, connection_info in connection_infos.items()
+        ], key=key_for_peer)
         for i in range(number_of_instances)
     }
-    return expected_received_values, received_values
+    return expected_peers_of_threads, peers_of_threads
